# backend/docs/swagger.yaml
basePath: /v1
definitions:
  dto.ActionResponse:
    properties:
      id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  dto.CreateProviderRequest:
    properties:
      channel:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      isDefault:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      secretConfig:
        additionalProperties: true
        type: object
      status:
        type: string
      type:
        type: string
    required:
    - channel
    - name
    type: object
...
  dto.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
      requestId:
        type: string
    type: object
...
  dto.ProviderResponse:
    properties:
      averageLatencyMs:
        type: integer
      channel:
        type: string
      config:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      isEnabled:
        type: boolean
      isPrimary:
        type: boolean
      lastError:
        type: string
      lastFailureAt:
        type: string
      lastSuccessAt:
        type: string
      name:
        type: string
      priority:
        type: integer
      status:
        type: string
      successRate:
        type: number
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ProviderTestRequest:
    properties:
      body:
        type: string
      channel:
        type: string
      dryRun:
        type: boolean
      recipient:
        type: string
      subject:
        type: string
    type: object
  dto.ProviderTestResponse:
    properties:
      channel:
        type: string
      checkedAt:
        type: string
      dryRun:
        type: boolean
      latencyMs:
        type: integer
      message:
        type: string
      providerId:
        type: string
      providerMessageId:
        type: string
      providerResponseSanitized:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SetDefaultProviderRequest:
    properties:
      isDefault:
        type: boolean
    type: object
...
  dto.ToggleProviderStatusRequest:
    properties:
      isEnabled:
        type: boolean
      status:
        type: string
    type: object
...
  dto.UpdateProviderRequest:
    properties:
      channel:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      isDefault:
        type: boolean
      isEnabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      secretConfig:
        additionalProperties: true
        type: object
      status:
        type: string
      type:
        type: string
    type: object
...
host: 127.0.0.1:9002
info:
  contact:
    email: support@minisource.com
    name: API Support
  description: Notification Service for Minisource - Handles Email, SMS, Push, and
    In-App notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Notifier Service API
  version: "2.0"
paths:
...
  /admin/providers:
    get:
      consumes:
      - application/json
      description: Retrieve all configured notification providers from providers table
      parameters:
      - description: Filter by channel (sms, email, push, webhook, in_app)
        in: query
        name: channel
        type: string
      - description: Filter by status (active, inactive, disabled, error)
        in: query
        name: status
        type: string
      - description: Filter by provider type (e.g., kavenegar, smtp)
        in: query
        name: providerType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProviderResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create a new notification provider
      parameters:
      - description: Provider data
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create provider
      tags:
      - Providers
  /admin/providers/{providerId}:
    delete:
      consumes:
      - application/json
      description: Delete (soft-delete) a provider
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete provider
      tags:
      - Providers
    get:
      consumes:
      - application/json
      description: Retrieve a single provider by ID
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get provider by ID
      tags:
      - Providers
    put:
      consumes:
      - application/json
      description: Update an existing provider
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      - description: Provider data
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update provider
      tags:
      - Providers
  /admin/providers/{providerId}/default:
    patch:
      consumes:
      - application/json
      description: Set or unset a provider as default for its channel
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      - description: Default status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetDefaultProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set provider as default
      tags:
      - Providers
  /admin/providers/{providerId}/status:
    patch:
      consumes:
      - application/json
      description: Enable or disable a provider, or update status field
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      - description: Status update
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.ToggleProviderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle provider status
      tags:
      - Providers
  /admin/providers/{providerId}/test:
    post:
      consumes:
      - application/json
      description: Test a specific provider connection (dry-run by default)
      parameters:
      - description: Provider ID
        in: path
        name: providerId
        required: true
        type: string
      - description: Test payload (optional, uses defaults if empty)
        in: body
        name: test
        schema:
          $ref: '#/definitions/dto.ProviderTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProviderTestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Test provider
      tags:
      - Providers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"